refactor: review compliance - rename useCallGuard to useNoInternetCallGuard
